SEQNG-449 Show the resume button when paused and viceversa
